Convert nightly release workflow to hub
Use hub CLI that is actually provided in the virtual environment.
Create the release as a draft, then after building and uploading
all assets, finalize the release with hub.
The release email should be sent at this point, listing all assets
added rather than only the source archives.

Properly delete the nightly tag from the repository
and recreate it on the current commit. In the workflow runs
so far, the nightly tag was pointing to an old commit,
which screwed up the release info.